+++ packages/core/include/combo/interface.h
@@ -8,6 +8,7 @@ void Interface_Update(GameState_Play* play);
 #if defined(GAME_MM)
 void Interface_NewDay(GameState_Play* play, int day);
 void Interface_SetHudVisibility(u16 hudVisibility);
+void Interface_SkipMoonCrash(GameState_Play* play);
 #endif
 
 #endif